prevent typedoc generating docs for private functions
